Refactor DataClass file having now one file per component
