+++ public_src/js/react/domainmodeleditor/DataClassFooter.js
@@ -1,0 +1,45 @@
+var React = require('react');
+
+var DataClassFooterComponent = React.createClass({
+    getInitialState: function() {
+        return {newname: ''}
+    },
+    handleChange: function(e) {
+        this.setState({newname: e.target.value});
+    },
+    handleAdd: function() {
+        if (this.props.handleAdd(this.state.newname)) {
+            this.setState({newname: ''});
+        }
+    },
+    handleEnterSubmit: function(e) {
+        if (e.keyCode == 13) {
+            this.handleAdd()
+        }
+    },
+    render: function() {
+        return (
+            <div className="panel-footer">
+                <div className="row">
+                    <div className="col-sm-12">
+                        <div className="input-group">
+                            <input
+                                type="text"
+                                className="form-control"
+                                placeholder="New attribute"
+                                value = {this.state.newname}
+                                onChange = {this.handleChange}
+                                onKeyDown = {this.handleEnterSubmit}
+                            />
+                            <div className="input-group-btn">
+                                <button className="btn btn-primary" onClick={this.handleAdd}>New attribute</button>
+                            </div>
+                        </div>
+                    </div>
+                </div>
+            </div>
+        )
+    }
+});
+
+module.exports = DataClassFooterComponent;
